new file for bq xlang tests
